Moved the Clock class to the shared folder
