+++ gameserver/src/main/java/dev/totallyspies/spydle/gameserver/game/GameLogic.java
@@ -1,5 +1,6 @@
 package dev.totallyspies.spydle.gameserver.game;
 
+import dev.totallyspies.spydle.shared.Clock;
 import dev.totallyspies.spydle.shared.model.ClientSession;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
